requirements.txt
streamlit
langchain
google-generativeai
mysql-connector-python
sqlalchemy
python-dotenv
